fix for int to unsigned int warnings.

+++ SYMPHONY/src/LP/lp_genfunc.c
@@ -2715,7 +2715,7 @@ int check_and_add_cgl_cuts(lp_prob *p, int generator, cut_data ***cuts,
       sym_cut->rhs = rhs;
       sym_cut->range = row_cut.range();
       //sym_cut->size = (num_elements * (int)((ISIZE + DSIZE) + DSIZE));
-      sym_cut->size = (DSIZE + num_elements * (int)((ISIZE + DSIZE)));
+      sym_cut->size = ((int)(DSIZE) + num_elements * (int)((ISIZE + DSIZE)));
       sym_cut->coef = (char *) malloc (sym_cut->size);
       sym_cut->sense = row_cut.sense();
       ((double *) (sym_cut->coef))[0] = 0; // otherwise valgrind complains.
